requirements.txt
cmd2~=2.4.3
pysocks
setuptools~=59.6.0